requirements.txt
Django
djangorestframework
psycopg2-binary
uuid
requests
pillow
pytest
pytest-django